using fastapi mechanisms
