+++ cr_scraper/ui/templates/grocery_list/archive_ui.html
@@ -1,10 +1,10 @@
 <div id="archive-ui" hx-target="this" hx-swap="outerHTML">
     {% if archiver.status() == "Waiting" %}
-    <button hx-post="/grocery_lists/archive/html">
+    <button hx-post="/grocery_lists/archive">
         Download Contact Archive
     </button>
     {% elif archiver.status() == "Running" %}
-    <div hx-get="/grocery_lists/archive/html" hx-trigger="load delay:500ms">
+    <div hx-get="/grocery_lists/archive" hx-trigger="load delay:500ms">
         Creating Archive... {{archiver.progress() * 100}}%
         <div class="progress">
             <div class="progress-bar" id="archive-progress" role="progressbar"
@@ -15,6 +15,6 @@
     {% elif archiver.status() == "Complete" %}
     <a hx-boost="false" href="/grocery_lists/archive/file" _="on load click() me">Archive Ready! Click here to download.
         &downarrow;</a>
-    <button hx-delete="/grocery_lists/archive/html">Clear Download</button>
+    <button hx-delete="/grocery_lists/archive">Clear Download</button>
     {% endif %}
 </div>
