+++ cr_scraper/ui/templates/grocery_list/rows.html
@@ -3,9 +3,9 @@
     <td><input type="checkbox" name="selected_list_ids" value="{{ list.id }}"></td>
     <td>{{ list.name }}</td>
     <td>{{list.groceries|length}}</td>
-    <td><a href="/grocery_lists/{{ list.id }}/edit/html">Edit</a></td>
-    <td><a href="/grocery_lists/{{ list.id }}/html">View</a></td>
-    <td><a href="#" hx-delete="/grocery_lists/{{ list.id }}/html" hx-target="closest tr" hx-swap="outerHTML swap:1s"
+    <td><a href="/grocery_lists/{{ list.id }}/edit">Edit</a></td>
+    <td><a href="/grocery_lists/{{ list.id }}">View</a></td>
+    <td><a href="#" hx-delete="/grocery_lists/{{ list.id }}" hx-target="closest tr" hx-swap="outerHTML swap:1s"
             hx-confirm="Are you sure you want to delete this shopping list?">Delete</a></td>
 </tr>
 {% endfor %}
