+++ cr_scraper/ui/templates/grocery_list/show.html
@@ -9,10 +9,10 @@
 </div>
 
 <p>
-    <a href="/grocery_lists/{{ grocery_list.id }}/add_recipe/html">Add recipe</a>
-    <a href="/grocery_lists/{{grocery_list.id}}/edit/html">Edit</a>
-    <a href="/grocery_lists/html">Back</a>
-    <button id="delete-btn" hx-delete="/grocery_lists/{{ grocery_list.id }}/html" hx-push-url="true"
+    <a href="/grocery_lists/{{ grocery_list.id }}/add_recipe">Add recipe</a>
+    <a href="/grocery_lists/{{grocery_list.id}}/edit">Edit</a>
+    <a href="/grocery_lists">Back</a>
+    <button id="delete-btn" hx-delete="/grocery_lists/{{ grocery_list.id }}" hx-push-url="true"
         hx-confirm="Are you sure you want to delete this contact?" hx-target="body">
         Delete
     </button>
