+++ cr_scraper/ui/templates/index.html
@@ -9,18 +9,18 @@
             <span class="error">{{ errors['name']}}</span>
             <br>
             <label for="shopping_list_recipe_url">Recipe url</label>
-            <input id="shopping_list_recipe_url" type="url" name="url" value="{{''}}" hx-post="/recipes/scrape/html"
+            <input id="shopping_list_recipe_url" type="url" name="url" value="{{''}}" hx-post="/recipes/scrape"
                 hx-target="next .error" hx-trigger="change, keyup delay:500ms changed" />
             <span class="error">{{ errors['url']}}</span>
             <br>
-            <button hx-post="/grocery_lists/new/html" hx-target="body">Create</button>
+            <button hx-post="/grocery_lists/new" hx-target="body">Create</button>
             <span hx-get="/grocery_lists/count" hx-trigger="revealed"><img id="spinner"
                     src="{{ url_for('static', path='/img/spinner.gif') }}" class="htmx-indicator" /></span>
         </fieldset>
     </form>
-    <input id="search" type="search" name="q" hx-get="/grocery_lists/html" hx-select="tbody > tr"
-        hx-target="#table_body" hx-indicator="#spinner" hx-push-url="true" hx-trigger="search, keyup delay:500ms"
-        value="{{q or ''}}" placeholder="Type list name here to search..." style="width: 50%;" />
+    <input id="search" type="search" name="q" hx-get="/grocery_lists" hx-select="tbody > tr" hx-target="#table_body"
+        hx-indicator="#spinner" hx-push-url="true" hx-trigger="search, keyup delay:500ms" value="{{q or ''}}"
+        placeholder="Type list name here to search..." style="width: 50%;" />
     <img id="spinner" class="htmx-indicator" src="{{ url_for('static', path='/img/spinner.gif') }}"
         alt="Request in flight..." />
 </div>
@@ -39,13 +39,13 @@
             <tr>
                 <td colspan="6" style="text-align: center;">
                     <span hx-target="closest tr" hx-swap="outerHTML" hx-select="tbody > tr" hx-trigger="revealed"
-                        hx-get="/grocery_lists/html?page={{ page+1 }}">Loading</span>
+                        hx-get="/grocery_lists?page={{ page+1 }}">Loading</span>
                 </td>
             </tr>
             {% endif %}
         </tbody>
     </table>
-    <button hx-delete="/grocery_lists/html" hx-confirm="Are you sure you want to delete these lists?" hx-target="body">
+    <button hx-delete="/grocery_lists" hx-confirm="Are you sure you want to delete these lists?" hx-target="body">
         Delete Selected
     </button>
 </form>
