+++ cr_scraper/ui/templates/layout.html
@@ -20,9 +20,9 @@
             </h1>
         </header>
         <div hx-boost="true">
-            <a href="/grocery_lists/html">Shopping lists</a>
-            <a href="/settings/html">Settings</a>
-            <a href="/help/html">Help</a>
+            <a href="/grocery_lists">Shopping lists</a>
+            <a href="/settings">Settings</a>
+            <a href="/help">Help</a>
         </div>
         {% block content %}
         {% endblock %}
